* Fixed warnings on compilers again.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@5654 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/common/utils.c
@@ -101,7 +101,7 @@ int strcasecmp(const char *arg1, const char *arg2)
   return (0);
 }
 
-int strncasecmp(const char *arg1, const char *arg2, int n)
+int strncasecmp(const char *arg1, const char *arg2, size_t n)
 {
   int chk, i;
